pyproject.toml
[tool.poetry]
name = "truck-food-finder"
version = "0.1.0"
description = ""
authors = ["Guilherme S.V. Beira <guilherme.vieira.beira@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
prettyconf = "^2.2.1"
django-rest-framework = "^0.1.0"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.1.0"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.1.0"
pytest-django = "^4.7.0"
django-coverage-plugin = "^3.1.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "food_truck_finder.settings"

[tool.coverage.run]
plugins = [ 'django_coverage_plugin', ]
omit = [
    "*/.venv/*",
    "wsgi.py",
    "asgi.py",
    "urls.py",
    "apps.py",
    "manage.py",
    ]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
